fix: Plan Get시 API Query를 쉽게 넣을 수 있도록 변경 #63
